+++ src/main/java/barch/mc_extended/Villagers/VillagerTrades.java
@@ -8,6 +8,8 @@ import net.minecraft.entity.Entity;
 import net.minecraft.item.ItemConvertible;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.registry.RegistryKeys;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.village.TradeOffer;
@@ -18,12 +20,15 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
 
+import static barch.mc_extended.MCExtended.MC_EXTENDED;
 import static barch.mc_extended.Registry.ModBlocks.*;
 import static barch.mc_extended.Registry.ModItems.*;
 
 public class VillagerTrades {
 
-
+    public static final RegistryKey<VillagerProfession> GEM_DEALER = of("gem_dealer");
+    public static final RegistryKey<VillagerProfession> BOTANIST = of("botanist");
+    public static final RegistryKey<VillagerProfession> MINER = of("miner");
 
     public static void RegisterTrades() {
 
@@ -36,7 +41,7 @@ public class VillagerTrades {
     public static void RegisterGemDealerTrades() {
 
         // novice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.GEM_DEALER, 1, factories -> {
+        TradeOfferHelper.registerVillagerOffers(GEM_DEALER, 1, factories -> {
             factories.add(new TradeOffers.SellItemFactory(Items.AMETHYST_SHARD, 1, 4, 1));
             factories.add(new TradeOffers.BuyItemFactory(Items.LAPIS_LAZULI, 4, 16, 1));
             factories.add(new ExplicitTradeFactory(Items.DIAMOND_ORE, 1, Items.COPPER_INGOT, 8, Items.DIAMOND, 3, 16, 2, 0));
@@ -45,7 +50,7 @@ public class VillagerTrades {
 
         });
         // apprentice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.GEM_DEALER, 2, factories -> {
+        TradeOfferHelper.registerVillagerOffers(GEM_DEALER, 2, factories -> {
             factories.add(new ExplicitTradeFactory(Items.DEEPSLATE_DIAMOND_ORE, 1, SILVER_INGOT, 6, Items.DIAMOND, 3, 16, 3, 0));
             factories.add(new ExplicitTradeFactory(Items.DEEPSLATE_LAPIS_ORE, 1, SILVER_INGOT, 3, Items.LAPIS_LAZULI, 9, 16, 2, 0));
             factories.add(new TradeOffers.BuyItemFactory(Items.GOLD_INGOT, 3, 16, 2));
@@ -54,7 +59,7 @@ public class VillagerTrades {
             factories.add(new ExplicitTradeFactory(Items.NETHER_QUARTZ_ORE, 1, Items.COPPER_INGOT, 2, Items.QUARTZ, 3, 16, 2, 0));
         }) ;
         // journeyman
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.GEM_DEALER, 3, factories -> {
+        TradeOfferHelper.registerVillagerOffers(GEM_DEALER, 3, factories -> {
             factories.add(new TradeOffers.SellItemFactory(Items.ECHO_SHARD, 12, 1, 10));
             factories.add(new ExplicitTradeFactory(Items.QUARTZ_BLOCK, 1, Items.COPPER_INGOT, 1, Items.QUARTZ, 4, 16, 3, 0));
             factories.add(new ExplicitTradeFactory(Items.CRYING_OBSIDIAN, 1, Items.EMERALD, 1, Items.GHAST_TEAR, 1, 16, 4, 0));
@@ -63,7 +68,7 @@ public class VillagerTrades {
             factories.add(new TradeOffers.SellItemFactory(Items.OBSIDIAN, 3, 1, 16,3));
         });
         // expert
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.GEM_DEALER, 4, factories -> {
+        TradeOfferHelper.registerVillagerOffers(GEM_DEALER, 4, factories -> {
             factories.add(new ExplicitTradeFactory(Items.RAW_COPPER, 1, Items.EMERALD, 1, Items.COPPER_INGOT, 2, 16, 4, 0));
             factories.add(new ExplicitTradeFactory(RAW_SILVER, 1, Items.GOLD_NUGGET, 2, SILVER_INGOT, 2, 16, 4, 0));
             factories.add(new ExplicitTradeFactory(Items.RAW_GOLD, 1, SILVER_NUGGET, 4, Items.GOLD_INGOT, 2, 16, 5, 0));
@@ -72,7 +77,7 @@ public class VillagerTrades {
             factories.add(new ExplicitTradeFactory(Items.EMERALD_ORE, 1, Items.COPPER_INGOT, 1, Items.EMERALD, 4, 16, 4, 0));
         });
         // master
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.GEM_DEALER, 5, factories -> {
+        TradeOfferHelper.registerVillagerOffers(GEM_DEALER, 5, factories -> {
             factories.add(new ExplicitTradeFactory(RUBY, 1, null, 0, Items.EMERALD, 15, 16, 5, 0));
             factories.add(new ExplicitTradeFactory(SAPPHIRE, 1, null, 0, Items.EMERALD, 15, 16, 5, 0));
             factories.add(new ExplicitTradeFactory(Items.RAW_COPPER_BLOCK, 1, Items.EMERALD, 1, Items.COPPER_INGOT, 13, 16, 5, 0));
@@ -83,14 +88,14 @@ public class VillagerTrades {
         });
 
         // gift
-        VillagerInteractionRegistries.registerGiftLootTable(ModProfessions.GEM_DEALER, Identifier.of("mc-extended", "gem_dealer"));
+        VillagerInteractionRegistries.registerGiftLootTable(GEM_DEALER, RegistryKey.of(RegistryKeys.LOOT_TABLE, Identifier.of("mc-extended", "gem_dealer")));
 
     }
 
     public static void  RegisterBotanistTrades() {
 
         // novice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.BOTANIST, 1, factories -> {
+        TradeOfferHelper.registerVillagerOffers(BOTANIST, 1, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Blocks.CHERRY_SAPLING, 6, 16, 3));
             factories.add(new TradeOffers.BuyItemFactory(Blocks.MANGROVE_PROPAGULE, 6, 16, 3));
             factories.add(new TradeOffers.SellItemFactory(Blocks.PINK_PETALS, 2, 12,16 ,4));
@@ -99,7 +104,7 @@ public class VillagerTrades {
 
         });
         // apprentice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.BOTANIST, 2, factories -> {
+        TradeOfferHelper.registerVillagerOffers(BOTANIST, 2, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Blocks.SPRUCE_SAPLING, 4, 16, 3));
             factories.add(new TradeOffers.BuyItemFactory(Blocks.JUNGLE_SAPLING, 4, 16, 3));
             factories.add(new TradeOffers.BuyItemFactory(Blocks.ACACIA_SAPLING, 4, 16, 3));
@@ -111,7 +116,7 @@ public class VillagerTrades {
 
         }) ;
         // journeyman
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.BOTANIST, 3, factories -> {
+        TradeOfferHelper.registerVillagerOffers(BOTANIST, 3, factories -> {
             factories.add(new TradeOffers.SellItemFactory(Items.SEAGRASS, 1, 8, 16,5));
             factories.add(new TradeOffers.SellItemFactory(Items.KELP, 1, 16, 16,5));
             factories.add(new TradeOffers.SellItemFactory(Items.SEA_PICKLE, 1, 4, 16,5));
@@ -133,7 +138,7 @@ public class VillagerTrades {
 
         });
         // expert
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.BOTANIST, 4, factories -> {
+        TradeOfferHelper.registerVillagerOffers(BOTANIST, 4, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.BEETROOT_SEEDS, 48, 16, 4));
             factories.add(new TradeOffers.BuyItemFactory(TOMATO_SEEDS, 48, 16, 4));
             factories.add(new TradeOffers.BuyItemFactory(Items.MELON_SEEDS, 18, 16, 5));
@@ -148,7 +153,7 @@ public class VillagerTrades {
             factories.add(new TradeOffers.SellItemFactory(Blocks.TALL_GRASS, 1, 1, 16, 6));
         });
         // master
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.BOTANIST, 5, factories -> {
+        TradeOfferHelper.registerVillagerOffers(BOTANIST, 5, factories -> {
             factories.add(new TradeOffers.SellItemFactory(Items.MOSS_BLOCK, 1, 12, 16, 5));
             factories.add(new TradeOffers.SellItemFactory(Items.CLAY_BALL, 1, 8, 16, 5));
             factories.add(new TradeOffers.SellItemFactory(Items.FLOWERING_AZALEA, 1, 8, 16, 5));
@@ -175,7 +180,7 @@ public class VillagerTrades {
     public static void RegisterMinerTrades() {
 
         // novice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 1, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 1, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.COBBLESTONE, 16, 16, 1));
             factories.add(new TradeOffers.BuyItemFactory(Items.IRON_PICKAXE, 1, 16, 3));
             factories.add(new TradeOffers.SellItemFactory(Items.COAL, 1, 8, 8, 1));
@@ -184,7 +189,7 @@ public class VillagerTrades {
 
         });
         // apprentice
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 2, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 2, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.STONE_SHOVEL, 1, 8, 4));
             factories.add(new TradeOffers.BuyItemFactory(Items.COAL, 16, 16, 2));
             factories.add(new TradeOffers.SellItemFactory(Items.DIORITE, 1, 16, 16, 3));
@@ -193,7 +198,7 @@ public class VillagerTrades {
 
         }) ;
         // journeyman
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 3, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 3, factories -> {
             factories.add(new TradeOffers.SellItemFactory(Items.DEEPSLATE, 1, 32, 16, 3));
             factories.add(new TradeOffers.SellItemFactory(Items.POINTED_DRIPSTONE, 1, 12, 8, 5));
             factories.add(new TradeOffers.SellItemFactory(Items.TUFF, 1, 24, 16, 4));
@@ -202,7 +207,7 @@ public class VillagerTrades {
 
         });
         // expert
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 4, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 4, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.SHEARS, 1, 12, 4));
             factories.add(new TradeOffers.BuyItemFactory(Items.DIAMOND_SHOVEL, 1, 8, 7, 4));
             factories.add(new TradeOffers.BuyItemFactory(Items.DIAMOND_HOE, 1, 8, 7, 8));
@@ -215,7 +220,7 @@ public class VillagerTrades {
 
         });
         // master
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 5, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 5, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.OAK_LOG, 4, 24, 4));
             factories.add(new TradeOffers.SellItemFactory(Items.DIAMOND, 3, 1, 8, 5));
             factories.add(new TradeOffers.SellItemFactory(RUBY, 24, 1, 6, 7));
@@ -232,7 +237,7 @@ public class VillagerTrades {
         // maybe when I am more competent I can add a level.
 
         // overboard
-        TradeOfferHelper.registerVillagerOffers(ModProfessions.MINER, 6, factories -> {
+        TradeOfferHelper.registerVillagerOffers(MINER, 6, factories -> {
             factories.add(new TradeOffers.BuyItemFactory(Items.DIAMOND_SWORD, 1, 12, 4));
             factories.add(new TradeOffers.SellItemFactory(Items.GLOW_INK_SAC, 3, 5, 10, 3));
             factories.add(new TradeOffers.SellItemFactory(Items.SCULK, 4, 1, 6, 3));
@@ -245,9 +250,8 @@ public class VillagerTrades {
             factories.add(new TradeOffers.SellItemFactory(Items.EXPERIENCE_BOTTLE, 3, 1, 8, 3));
         });
 
-
         // gift
-        VillagerInteractionRegistries.registerGiftLootTable(ModProfessions.MINER, Identifier.of("minecraft", "miner"));
+        VillagerInteractionRegistries.registerGiftLootTable(GEM_DEALER, RegistryKey.of(RegistryKeys.LOOT_TABLE, Identifier.of("mc-extended", "miner")));
 
     }
 
@@ -287,6 +291,9 @@ public class VillagerTrades {
                     this.priceMultiplier
             );
         }
+    }
+    private static RegistryKey<VillagerProfession> of(String id) {
+        return RegistryKey.of(RegistryKeys.VILLAGER_PROFESSION, Identifier.of(MC_EXTENDED, id));
     }
 
 }